requirements.txt
avro==1.12.0
confluent-kafka==2.5.0
cramjam==2.8.3
fastparquet==2024.5.0
fsspec==2024.6.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pyarrow==17.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
watchdog==4.0.1
